requirements.txt
flask
pymongo
google-genai
python-dotenv
gunicorn
requests